🐛 fix(docker): 修复 Agent Docker 部署和数据库类型不匹配问题
- 修复 Agent Dockerfile 缺少 server Cargo.toml 导致 workspace 构建失败
- 修复 verify_device_token 中 i64/i32 类型不匹配（PostgreSQL SELECT 1 返回 INT4）
- 增强 WebSocket handler 错误日志输出完整错误链

+++ crates/server/src/db/agent_ops.rs
@@ -103,7 +103,7 @@ pub async fn connect_device(pool: &PgPool, device_id: Uuid) -> Result<DeviceConn
 /// - Ok(false): 验证失败
 /// - Err: 数据库错误
 pub async fn verify_device_token(pool: &PgPool, device_id: Uuid, auth_token: &str) -> Result<bool> {
-    let result: Option<(i64,)> = sqlx::query_as(
+    let result: Option<(i32,)> = sqlx::query_as(
         r#"
         SELECT 1 FROM devices WHERE device_id = $1 AND auth_token = $2
         "#,
